- extract file repository from test
